+++ CHANGELOG.md
@@ -1,6 +1,25 @@
 # Changelog 
 All notable changes to this project will be documented in this file.
 
+## 3.2.1 (24/08/2020)
+* Fix problem with threads in ActionManager with method clean() and usage in BasicHook
+* Add dismissMobileDriver in DriverManager
+
+## 3.2.0 (10/08/2020)
+* LocatorManager refactor
+  - catch exceptions
+  - split and find locator file (max 1 folder in locators folder)
+  - getProperty method, if is another locator file loaded, load the requested
+* Add LocatorTypesEnum
+* Add getName() method in ProjectTypeEnum with name of driver instance + call it in DriverManager.getName()
+* Method loadLocators in the constructor PageBase and getLocator that returns By searching in LocatorTypesEnum
+* Refactor PageBase, PageBaseMobile and PageBaseWeb in methods to support parameter String locatorName and deprecated in old methods.
+* Replace in places that were using getRootLogger() for Logger.getLogger(class).
+* Api fix in CommonSteps (Delete ':')
+* Remove unused methods in Utils class.
+* Add ActionManager, WebActionManager and MobileActionManager with all static methods and using locator manager. (This replaces PageBase).
+* Fix fluent time in Constants and remove constants without use.
+
 ## 3.1.2 (04/08/2020)
 * Move dragAndDrop method from PageBase to PageBaseWeb: these methods is specific for web project.
 * Add waitForElementClickable to PageBase.